Fix another few tricky log bugs.
This is becoming a quagmire. The problem is that appending to the log may
cause objects to move around when you don't expect it, since the cleaner
has been triggered in the meantime in order to satisfy the write.

For example, the following is dangerous:
    old_obj = hash->lookup(key)
    log->append(new_obj)
    hash->put(key, new_obj)
    log->free(old_obj)

Why? Because old_obj may have been relocated when appending new_obj to
the log. The pointer is invalidated. In general, when appending to the
log, any pointers you have into the log may become invalidated.

Pretty sweet, huh? Time to rethink this...