Add make docs using doxygen
There are currently no documentation warnings, and I'd like to keep it that way
in the future. Something that is entirely undocumented will not produce a
warning (for now).

I'd recommend everyone to symlink hooks/pre-commit-docs to
.git/hooks/pre-commit so that their clients will not allow them to commit if
they introduce documentation warnings. If this becomes a problem, we can add a
similar server-side hook to reject pushes that introduce documentation
warnings.