Change conditional operations to be more general
Now the read, write, and delete operations have shared, general rules for when
to reject an operation. The idea is that commonly used rules will be exposed to
the user with friendly names in the high-level language bindings, and the
general-purpose interface will be available for those occasional cases when
they're useful. See the documentation for rcrpc_reject_rules for more info.