Change delete semantics for returned version number
The version number returned by delete is now the version number that reflects
the current state of the object. If the operation succeeded (the object no
longer exists), that is RCRPC_VERSION_NONE. If it failed, the behavior is the
same as before: it returns the version of the object that caused the operation
to fail.