BackupServer::getSegmentMetadata is working
Unit testing for BackupServer is now relatively complete.  Some
more aggressive checks are waiting on a slightly more realistic
implemenation of the backup and recovery protocols.  There is
a bug in the Bitmap class preventing some testing as well which
would be a good target for me to fix up next.