Change server to bump version number by 1 on updates
On updates, the server will now increase the version number of an object by
exactly 1. The master vector clock behavior is now documented in Server.h.

Here's why it's useful for client-side transactions:
Once a minitransaction commits, we want to be able to report the version
numbers resulting from any writes that took place in the MT. However, other
clients are allowed to "help out" by racing the client executing the MT to
write back the new blobs to their final destinations. When this happens, the
client executing the MT will not be the one to do the final writes on all of
the objects and thus can't get their new version numbers that way. With this
new server guarantee, the client executing the MT knows that the final version
of each object will be the version at which the mask was added plus 1.