Fix HashTable::PerfDistribution::bins initialization
The unit tests caught this bug (!) when I was messing with something
else entirely.

sizeof(bins) is no longer the size of the array, now that bins is just a
pointer to a heap-allocated piece of memory. Therefore, memset()ing
sizeof(bins) to 0 is no longer valid. Only the first entry of bins was
being cleared out, but subsequent entries could have had garbage.

This bug was introduced in the following commit:

commit 6497a0fbbc376b35dda7e6df1fe14055202d13af
Author: Diego Ongaro <ongaro@cs.stanford.edu>
Date:   Mon May 3 14:34:39 2010 -0700

    hashtable: Change PerfDistribution::bins to heap allocation

    This fixes the segfault at ServerMain.cc's main() from growing the stack
    too quickly.

I also fixed a typo in the comment for bins.