Add missing Doxygen \file comments
This is mostly trivial, but there was one important case:
TCPTransport.cc had a file comment, but it wasn't labeled as \file.
Because the next documentable thing after the comment was the RAMCloud
namespace (as is commonly the case), that comment ended up applying to
the entire RAMCloud namespace and not the TCPTransport.cc file.