Move reusable code out of RamCloudClient
Refactoring common code out of RamCloudClient's methods had the nice
side effect of reducing the number and complexity of its unit tests.

Client will be used as the base class of other RPC proxy classes (such
as BackupClient) in the future, and the changes to Rpc.h will also allow
for some refactoring in Server.cc.