Separate MasterClient from RamCloud class
There's a lot here, but many of the changes are renames. A couple of
notes:

The client binary now takes a coordinator locator, not a master locator.
It queries the coordinator to find a master (though not the tablet map,
just the server list for now).

MasterServer and MasterClient now share BindTransport-based unit tests.