Improve BindTransport
BindTransport now uses Server::handleRpc rather than Server::dispatch
because it's handleRpc's job to catch ClientExceptions that the Server
handler throws.

Additionally, the Server argument to BindTransport can now be filled in
later. This is desirable if the Server's constructor calls any methods
on TransportManager that depend on the BindTransport being registered.