Add support to BindTransport to talk to multiple servers.
A single server can be added as usual at the mock: address, but
multiple hosts can be registered as well by calling addServer()
and creating a fictitious service locator in the form
mock:host=myserver.  Once a server is registered it can be
contacted by passing the service locator through the transportManager
in the usual way (getSession()).