Pipework for the per recovery master portion of recovery.
Adds MasterClient::recover, MasterServer::recover,
BackupManager::recover, and MasterServer::recoverSegment.

MasterServer::recoverSegment needs to be filled in still with code
to replay filtered log segments, one per call.  All Coordinator
recovery routines still need to be layered on this to kick off
this recover procedure on the recovery masters.

There is also likely a serious problem in that Masters perform
recovery as a single RPC to the Coordinator and are, therefore,
blocked to all calls until the process completes.