Improve makefile performance
The list of files for cpplint was being generated every time make was
run. This invoked pragmas.py for every source file, regardless of the
make target.

Now, the list of files for cpplint is only generated during make check.
Moreover, I've added a filter mode to pragmas.py so we only have to
invoke it (and the Python interpreter) once to filter a list of files.