Added naive config file support to OptionParser.
Be forewarned: it's easy to accidentally change argument defaults
such that testing scripts won't run as expected (setting replicas
to non-zero, for example).

To override any commandline options create a .ramcloud file in
the directory you run your RAMCloud binaries from.  The format
for the file is a newline separated list of key, values separated
by an equal.  The keys correspond to the long options of the
programs as output by help.

Keys that aren't understood by a particular binary are ignored.

The file from which config options is read can be overridden on
the commandline via -c.

.ramcloud example:
logLevel=9
coordinator=tcp:host=127.0.0.1,port=12345
replicas=2