infrc: Fix async RPCs, etc
I added a nonce in a header to all RPC requests and responses, since the
assumption that a QP uniquely identifies an RPC does not hold with the
async API.

The receive buffers in clientSrq are now reserved at the point when an
RPC is sent, and I added a clientSendQueue for RPCs that can not be sent
because their responses may not have a receive buffer available. If a
response is received and the receive buffers are running low, the
response is copied and the receive buffer is returned immediately.

I split the old srq into a clientSrq and a serverSrq so that servers
that also act as clients can dedicate receive buffers to their client
RPCs.

I moved to a common client receive CQ rather than having one per QP so
that receive buffers can be returned to clientSrq quickly when they're
running low.

I changed the number of buffers per SRQ from 4 to 8 so that we can
support up to 4 simultaneous RPCs very efficiently, 4 more with a copy
of their response, and any more will be queued.

infrc now passes transport-smack.