Improve Crc32C
Crc32C is now a class so that we can preset the value to -1 and invert the
result, which is supposed to detect leading and trailing bits better.

The unit tests for Crc32C now use a value parameterized tests (using gtest
rather than CPPUNIT) to test both the software and the hardware implementation,
where available.