Add an analogue to CycleCounter called TimeCounter, which uses clock_gettime and reports the time in nanoseconds. This is useful for measuring timeout periods, etc.
This is almost identical to CycleCounter, so perhaps some
refactoring should be done so we can have one class with a
selectable time source, or suitable some hierarchy.

The reason I'm not using rdtsc() is because I don't need extreme
granularity and didn't want to worry about lack of rdtsc()
synchronisation across cores.