Finish plumbing type changes.
The more I think about this, the more I dislike it. This is especially
nasty because the stack on AMD64 grows down from 2^47 - 1, so if we
start allocating more bits for a type field we're restricting legit
addresses. Yeah, they're stack addresses, but that still kinda sucks.

If we decide to write tombstones to the log on recovery, we can just
use the Log's type information for this purpose. However, we're not
doing that now to avoid extra backup traffic and so we need a way
to differentiate types...