Changes to make fast+infud work.
Infiniband doesn't have a notion of static ports. I'm not even
sure if there are simple ways to get static addresses.

We avoid this issue with InfRCTransport by bootstrapping over
IP/UDP and using known addresses/ports. However, with the UD
driver, there's no concept of connections.

This presents two issues. First, we need to handle dynamic
ServiceLocators that get updated when Transports/Drivers are
configured. My solution was to add a getServiceLocator() method
to all Transports (and underlying Drivers). These may be updated
if dynamic addresses are allocated. TransportManager can then
be queried to get an updated ServiceLocator string, which it can
then use to enlist with the Coordinator.

Second, we need a way to resolve these dynamic addresses.
Fortunately, the Coordinator does the latter for us already.

To use a fast+infud RAMCloud, we need now just ensure that the
Coordinator uses both fast+infud and another statically addressed
transport. The rest of the nodes may use dynamic addresses and
fast+infud only.

This could be extended to other Transports to avoid static port
allocation all over the place, but I've not modified them.