Import post-receive-jira hook in case fiz dies
This is a server-side git hook that scans the messages of new commits on
the master branch and updates Jira issues accordingly. There are two
things you can do right now:
1. If your commit message says something like "Closes RAM-153" (variants
of "close", "fix", and "resolve" are allowed here; case-insensitive),
the commit message will be logged to the issue, and the issue will be
closed.
2. If your commit message otherwise mentions an issue number like
"RAM-153" (case-insensitive), the commit message will be logged to the
issue.