Overhaul InfAddress class and move it inside Infiniband class
The most important change here is that Infiniband address handles are
now created on demand (lazily). This makes a big performance difference
because the cost of an ibv_create_ah call is fairly high: mlx4_create_ah
calls into the kernel driver every time to figure out if the physical
port on the device is an Infiniband port or an Ethernet port. There's
obviously some room for trivial optimization inside libmlx4 here, but
this change hides most of the overhead.