Rework the way partitions are calculated within the TabletProfiler. The goal is to return the minimum and maximum number of bytes and referants that could be within the partitions generated.
Instead of attributing parent counts to the first leaf bucket, we now
mostly ignore the parents until we decide to make a partition boundary.
When we do so, we know that the partition has at least as many bytes
and referants as were accounted for in the leaf buckets, and at most
that number plus the amount of bytes that could have come from parent
buckets leading to each boundary.

We currently decide to split at the max, which ensures that we never
exceed what we want, though we could be off by 2 * N-1 * SPLIT_BYTES,
which N is the maximum number of levels. For current parameters,
that means we could have ~100MB less than our maximum target, but
that's probably fine.