Remove 'checksum' and 'data_len' from Object. The former was never used and should probably be in SegmentEntry for all things written to the Log instead. The latter was duplicated in the SegmentEntry. Both also could have been 32-bit, rather than 64-bit.
This saves 16 bytes per Object (total is now 32, rather than 48,
including SegmentEntry). I'll probably take 4 back for the
checksums, but we can shrink other fields. E.g., SegmentEntry.type
needn't be 32-bit, especially if we have a crc that can be verified
if we are in disaster recovery mode and are trying to scrape out
entries from random byte arrays.