Add mechanisms for synching the MasterServer's Will with the Coordinator. There are two ways this happens:     1) With the TABLETS_RECOVERED rpc at the end of recovery.     2) With a new SET_WILL rpc at any other point.
When tables are created, nothing new happens. The Coordinator just
adds the empty tablet to the Will it has.

The next step is to update the Coordinator when a master's set of
partitions starts to become imbalanced. The most straightforward
way is to just do it periodically. A slightly smarter method is to
track changes and periodically check if a new Will is necessary.