scripts/recovery: Drop umask and ulimit
The umask's original purpose was to create 0666 backup files in /var/tmp/ so
that other users could overwrite them. This never seemed to do the trick, and
we no longer place backup logs there.

The ulimit's purpose is to enable core dumps. Unfortunately, these are
gigabytes in size and are written over NFS, so most of the time you don't want
these. The ulimit can be reenabled by setting the remoteprecmd variable.