Replace sleep statements in script/recovery with new ensureHosts command
The goal is to make recovery more reliable. Rather than assume all servers are
online after sleeping an arbitrary amount, the new ensureHosts binary asks the
coordinator for a list of servers that have registered. It will then wait until
the correct number of servers are online (up to a timeout).

A large part of this patch is delaying when masters enlist until
MasterServer::run(). The problem before was that they would enlist with the
coordinator well before they were ready to serve RPCs. If the client was then
quick enough to create a table, the coordinator would attempt to open a session
with that master while it was initializing its log memory. Because the master
was not servicing its transport during that time, opening that session would
time out.