Removed n^2 algorithm from recovery master routines.
Whenever a getRecoveryData completed successfully the recovery master
would walk the remaining segment backup locations to find other backups
for the same segment and mark them off.  This prevented them from being
requested later.

Recovery masters now build an unordered_multimap of all the entries
in the segment backup location table and refer to that to mark segments
with the same segmentId off after a successful replay.

This very clearly reduces the trend of recovery time from polynomial
to linear when many many segments are being recovered as is the
case as the number of partitions scales out.