Fix TransportManager regressions
Most of these fixes are adding Dispatch::Lock to code that can be called
from worker threads.

The other one is querying the transport for its listening locator rather
than using the one provided on the command line. This is required for
infud to work, which generates its own address. It was broken in 89a6fa.