StringKeyAdapter supports a string key style data model.
The current implementation takes shortcuts on hash collisions
in order to be simple and fast.  Though, it uses a collision
resistant hash which seems to be quite effective.

A simple benchmark yields about 6.3 us RTT for reads of small
objects with short (a few byte) keys.

This patch also augments the RamCloud client to support writes
of non-contiguous memory chunks from the application without
copying.