Rework how TabletProfiler data is maintained. The threaded log cleaner make this somewhat more difficult, as we use the LogTime (<SegmentId,Offset>) to determine which buckets in the structure to decrement when untracking. However, this requires us to ensure that tracking ops are done in log order. Yet cleaning disturbs the order (cleaned-to Segments have higher IDs than the head).
So, instead, make this all asynchronous. Now the cleaner, before
considering a Segment for cleaning, iterates over it and fires a
"scanning" callback on all entries. This callback is presently used
to track the objects and tombstones added to the log. Each Segment
is scanned only once and in exact order of Segment Id. Entries that
have been relocated due to cleaning will be scanned again in their
new Segments. Thus we untrack in the relocation callback and track
again in the scanning callback.

The "scanning" name sucks, so feel free to suggest alternatives.