Add ServerRpcPool, the sole and central allocator for ServerRpc- derived objects. This object stamps all incoming RPCs with an epoch value and tracks all outstanding ServerRpcs in the system. We can use this functionality to grab the current epoch and check when all RPCs as old or older have left the system. The cleaner will use this to know when its safe to return cleaned segments to the free list.
Note that this doesn't handle ClientRpcs, so if any ClientRpcs
are initiated by the server without being prompted by an incoming
ServerRpc (or the action is taken after the ServerRpc is replied
to), then this won't help. We don't have any such situations yet,
so far as I'm aware.

Also, toss the old boosty ObjectPool and replace with a trivial
reimplementation. The interface is unchanged except that it does
not automatically clean up for you (but it will assert if you
don't destroy all your objects before destroying the pool).