Cleaner optimisation: when writing out an object, try placing it in a previous segment first. This allows us to pack our segments better.
E.g. consider if we clean a 1MB object but have 2^20 - 1 bytes left
in the new segment we're writing to. We have to allocate a new
segment to write the object, but in the future we could pack smaller
objects into that 2^20 - 1 bytes. Before we were ignoring that empty
space. Now we don't.