Move process globals into new Context class
Closes RAM-212: Destruction of transportManager isn't safe
                (ordering issues).

From the class-level docs:
Context is a container for global variables.

Its main purpose is to allow multiple instances of these variables to
coexist in the same address space. This is useful for a variety of
reasons, for example:
 - Accessing RAMCloud from multiple threads in a multi-threaded client
   application without locking.
 - Running a simulator of RAMCloud in which multiple clients and servers
   share a single process.
 - Running unit tests in parallel in the same process.

Context also defines an explicit order in which these variables are
constructed and destroyed. Without such an ordering, it's easy to run
into memory corruption problems (e.g., RAM-212).

The current context is stored in thread-local state. Each entry points
into the RAMCloud library and the main functions of daemons and threads
set current the context using a Context::Guard. The remainder of
RAMCloud code uses Context::get() when it needs to access variables in
the current context.