Update some BackupManager docs.
From Diego; part of a BackupManager cleanup that has suffered some
bit rot.

This patch removes the hack to BackupManager to allow it to use
uniform random distribution of segments which breaks the old
recovery_dist.py CDF generation script.  Rather than perpetually
fixup a script we don't use for all changes it has been removed
feel free to resurrect it if we need it at some point.  It'll
be safer just to repair it then.