Add a ClusterPerf test for testing basic connectivity.
./scripts/clusterperf.py readAllToAll creates a Master with a
single object stored in it on every host in the cluster and then
starts a client on each host in the cluster having each client,
in turn, read the one object from each of the Masters in the
cluster.

This is good for sanity checking to make sure every host pair
in the cluster can send and receive RPCs from one another.