ReplicaManager nows supports concurrent access from multiple threads.
The ReplicaManager and ReplicatedSegments shared an internal giant
lock that serializes all access to their shared state.  Right now
calling sync() on a segment may spin waiting for another thread to
exit the proceed loop at which point the waiting thread may find
its data has already been synced for it.  The current implemenation
can exhaust the cores with spining threads, though this will be
remedied in a future commit.