Cache more detailed information about each server in ServerTrackers.
This allows getRandomServerIdWithService to find services without
acquiring locks as well as a lock-free getSession.

ServerTypeMask is growing tentacles, so we'll probably want to
factor that into a separate, easy-to-use class.  Also, we probably
need to think through ServerList/Tracker session acquisition
semantics in order to provide the simplest possible failure handling
for clients.  In particular, these classes may want to be aware or
generate hintServerDown calls in order to prevent flooding the
coordinator.