Add monitor which notifies ReplicaManager of backup failures.
Internally, the BackupFailureMonitor uses a ServerTracker to listen
for changes to the master's main ServerList, sifts through those
changes, and calls into the ReplicaManager to take corrective
actions when one of those changes is a backup failure.  This
BackupFailureMonitor runs in a separate thread this allows it
to react to failures in a timely manner and provides a context
to ensure that all corrective actions have been completed.