Read replicas from disk when backups start.
Rather rudimentary support that doesn't address several issues.  Garbage
collection of stored replicas that aren't needed is not addressed.  It
also does not recover the length of the segments for segments which were
open during a crash.  This is used in recovery to determine which copy
of a segment should be considered the real log head (though that should
change shortly with the new setMinOpenSegmentId functionality).

This patch doesn't distinguish between replicas found open on storage
and replicas that were opened the usual way.  It would probably be best
to have a special distinction for replicas which were open when they
crashed, but are no longer open since the original backup they were on
crashed.

Some older code expects that any open segment resides in memory.  In
order to keep from breaking that code stored replicas which were found
open on storage are loaded into RAM.

Finally, this code doesn't augment backups to inject a checksum for open
backups when a backup is exiting or under duress (which will also
require us to augment checksum to cover the footer log entry type).