Atomic segment replication protocol.
This adds a new flag to backup write rpcs which sets an open replica
into an atomic write mode where either the entire replica commits
on close or the replica is guaranteed not to be used.

This will be important in backup recovery since we don't want
re-replication of closed segments to be mistaken for open replicas
particularly since they may contain a log digest and could be
mistaken for the head of the log during master recovery.