Enlistment now allows a server to supersede another.
This allows a backup which is re-enlisting with persistent data
to ensure that its predecessor instance has been removed from
the cluster before it is enlisted.  This ensures that servers
are guaranteed to see a failure notification for the old server
before the addition notification of the replacement instance.

For now backups don't persist their former id so this is unused,
but it lays the foundation for it in a future patch.

This also implies that enlistment (in addition to
hintServerDown) can cause the start of master recoveries.