Add CRASHED status to servers in server lists.
Servers in server lists have a status of UP, CRASHED, or DOWN.
This allows other servers in the cluster to take action on
server failure while knowing which servers they may need to
hold on to resources (replicas) for.  If a server is CRASHED
then any replicas for it need to be held by backups.

Trackers now receive notifications CRASHED events in addition
to the original ADDED and REMOVED events.

This patch is a small step toward making replica garbage collection
work, which is yet another small step toward cold start and
continuous recovery.