Make RAMCloud compile with newer Ubuntu/Debian toolchains.
New changes to Debian distros' toolchains trims unused symbols from
libraries even if they are explicitly linked against.  The linker
works left-to-right on the command line options.  If some symbol is
unused when the library is encountered in the list it is not included in
the output.  This requires the .o file being compiled to be listed first
(since it represents the top-level of symbols that must be included in
 the output).

The details are discussed here:
https://wiki.ubuntu.com/NattyNarwhal/ToolchainTransition

Adapted from a patch supplied by Markus Pilman <mpilman@inf.ethz.ch>.