Added serverSpan option to createTable
This patch enables the spanning of a single table across multiple servers. Upon
table creation, the variable "serverSpan" (which has a default value of 1)
defines the number of servers. The spanning is done by splitting the key-range
according to the specified number of servers. If the chosen serverSpan is
greater than the number of available servers then a single server gets to host
multiple key-ranges of a single table.