Add a GetHeadOfLog RPC to the master that returns the current <SegmentId, ByteOffset> of the head of the log. When creating tablets, store this value in the tablet map and will. During recovery, ignore objects that existed prior to this offset. This avoids resurrecting old objects if a tablet is migrated away and back (or if we someday support re-creating a tablet with the same ID) and a crash occurs.
The cleaner requires some special attention because it fills
in segments that have higher IDs than the current log head. This
means that old objects could appear ahead of the recorded log
head if a GetHeadOfLog RPC came in during cleaning. To fix this
we stamp each segment's header with the ID of the log head when
cleaning began. This guarantees that the survivor segment that
was generated contains data from only segments prior to the head
ID reported.

This mostly closes RAM-135. There remains a tiny issue where we
need to be careful that the hash table is purged of old objects
before accepting a tablet back again. The cleaner will currently
keep an object if the tablet exists and its in the hash table.
Unless we want to version tablets or be very careful to check
an object's position in the log before acting on it, this logic
could yield false positives if an old object is seen after the
tablet has migrated back. Purging the hash table is probably the
simplest option. Whether we do it on removal of a tablet, before
migration completes, or only on re-acceptance of a prior tablet
are all possible alternatives with different trade-offs.