Servers store a superblock to help manage cluster membership.
This patch is part of a series to provide efficient and safe garbage
collection of segment replicas from storage even across backup
failures.

There are two key features that the backup superblock provides:
1) Backups are now provided with a "cluster name" on startup that
prevents them from a) reusing replicas on storage that were written
by a backup operating under a different cluster name, and b) prevents
future backups from reusing the replicas that are written by the
starting backup unless their cluster name matches.  This prevents,
for example, one from running ./scripts/recovery.py and accidentally
having replicas reused by subsequent runs of the script.
2) The superblock contains the server id that formerly owned the replicas
found on disk. This is reported to the coordinator on enlistment. The
coordinator uses this to guarantee that all servers hear about the
failure of the former server id before hearing about the addition of
the server which is now reusing the former server's replicas.