Added a server-sided increment function
Increments a numeric object by a specifiable value. The
object should be an 8-byte, two's complement, little-endian integer.
If the object has a different size than 8bytes,
the method throws a STATUS_INVALID_OBJECT exception.The increment
value can be negative in order to decrease the value of the object.