Modifying clusterperf to print key lengths and improving formatting.
Clusterperf now prints out the key length of the object being
written or read along with the data size, since the performance
also depends on the key length. This closes RAM-394.

The read and write performance measurements with varying key
lenghts are now printed as a table for better readability.