Backups garbage collect replicas that may not get freed by masters.
When a backup fails a master replicates affected segments elsewhere.  In
doing so the master is released from the obligation of freeing the
storage of "lost" replicas since the backup may never return to service.

Now, backups periodically try to make incremental progress is freeing up
replicas stored on them. Most replicas are handled the usual way when
there aren't failures; the master frees replicas explicitly from the
backups. For the cases where the master has failed or a backup has
failed and restarted the garbage collector decides what to do.

The garbage collection routine also obviates the formerly planned "end
of master recovery" where backups were supposed to collectively free the
log of the crashed master.