Rework cluster.py to allow scripts which mess with the cluster as it runs.
The new Cluster class provides a way to start a group of RAMCloud
processes on the cluster which are cleaned up when the experiment is
over. Cluster allows servers to be added and killed over time. It also
allows arbitrary client binaries to be run across the cluster.

cluster.run() still exists for running simple, one-shot experiments
with a specific cluster configuration, but it is rewritten in terms of
this new class.