Added tabletSplit
I implemented tabletSplit which allows splitting a specific tablet into two
pieces. In a nutshell, I added a tabletSplit function to the coordinator and
to the master. A client calls tabletSplit on the coordinator and provides
the table name, the start and end hash key from the to be split tablet in
order to identify it and the hash key where to split. Then the coordinator
updates its tabletMap accordingly and calls the splitTablet function on the
respective master.