Added MasterService::getServerStatistics
I added a protobuf ServerStatistics with the intention to use it for
transferring all relevant statistical information about a single
master server to a client e.g. a workload manager. As of now, this
protobuf contains a nested message TabletEntry which gets instantiated
for every tablet that a master currently has (I assume the protobuf
will be enriched over time with fields for all relevant statistical
information). This TabletEntry is used to store the current number of
reads and writes to a single tablet. The update of this value happens
with no additional overhead besides the value incrementation itself.
I checked that via clusterperf.py, see https://gist.github.com/2494499.
The ServerStatistics protobuf gets exposed in its entirety
via a GetServerStatisticsRpc.