Factor out coordinator protobuf-based tablet map abuse.
Introduces TabletMap which houses the authoritative info on
Tablet placement. TabletMap is a rough-draft of what will
eventually be needed, but it consolidates logic and provides
basic thread-safety (as the coordinator is moving toward being
concurrent). TabletMap is serializable to the standard
ProtoBuf::Tablets used before for easy transmission to
other hosts.