Enumerate a table with once only semantics.
Adds an Enumerate RPC for retrieving part of a tablet from a
master. At the client library level, this RPC is wrapped in a loop to
enumerate the entire table. To resume in between iterations of the
loop, an iterator value is passed between the client and server.

This version of the patch supports once only semantics. That is, every
object that survives the entire lifetime of the enumeration is
guaranteed to be returned exactly one time. Any object that is only
alive during part of the enumeration (e.g. because it was deleted part
way through) is guaranteed to be returned exactly zero or one time.