First step toward CoordinatorServerList thread safety.
Provides monitor-style locking on operations to provide atomicity.
Many uses of the list span multiple operations and have varying
consistency needs; they will still need to be rethought on an
individual basis. Most of the code churn resulted from moving
to by-value semantics for access to list elements causing rework
of many of the users of the list as well as the addition of
methods to the list to update entries.