Recovery masters safely give up on recoveries aren't going well.
When a recovery master can't find a replica for some segment
for whatever reason (backups crashed since start of recovery,
checksum failure, etc.) it can now abort its portion of recovery.
It reports this failure to the coordinator which will restart a
followup recovery to recover the missing partition of the will
later.

Currently, to abort the recovery, the recovery master scans its
hashtable removing (from the hashtable and log) objects which
were part of the recovery. For now this is done synchronously
after notifying the coordinator and on the sole master worker.