Start of system test framework.
Scripts and runs system tests across the cluster in accordance with
the user's localconfig.py.
Includes a few very small tests for now. One tests a simple recovery
of one object. Others test recovery when recovery masters fail.

This patch also includes a way to push configuration options to
the coordinator. For now it is used just to set testing options
(that is, to ask the coordinator to crash some recovery masters).