Don't acquire BackupFailureMonitor lock on trackerChangesEnqueued.
This is serious and subtle ramifications. Acquiring the lock in
trackerChangesEnqueued would destablize the entire cluster when
handling failures. In a nutshell, sometimes the server would
already be in the process of recovering from backup failures
when an UPDATE_SERVER_LIST rpc came in. In this case
BackupFailureMonitor::trackerChangesEnqueued would be called
synchronous before the coordinator's rpc finished. However, since
that acquired the lock it had to wait on the handling of backup
failures. Unfortunately, handling backup failures sometimes
requires sending a setMinOpenSegmentId rpc to the coordinator
which would cause deadlock. This deadlock would break when the
coordinator timed out its update server list rpc, but since
the server couldn't process other membership service requests
in the meantime other servers may have timed out trying to
open a session to it (because opening a session requires a
getServerId rpc). All this lead to other servers trying to
evict the slow server which lead to more cluster membership
updates and more probabilty of the problem bringing down
other hosts.

Welcome to hell.