Fix racy deadlock which occurred sometimes on recovery masters.
The log.sync() at the end of recovery was acquiring a lock on the
head segment and then calling sync() on the underlying
ReplicatedSegment. In the meantime, if underlying replicatedSegment
was notified that it lost a replica would begin re-replication.
This would lead to deadlock.

In this case log.sync() cannot return until the segment finishes
recovering from lost replicas and the recovery from lost replicas
cannot complete because it requires the lock on the segment head
log.sync() is holding.