ReplicaManager queries Segments in preparation for new log module.
ReplicatedSegment::sync() calls back into Segment to determine how
much data can be replicated. ReplicaManager no longer holds a pointer
to the backing memory for each segment; segments are free to implement
storage however then choose. When a write rpc generated by
ReplicaMananger needs data from a segment it calls back into the
Segment providing it with the details needed for the data to be
appended into the rpc's buffer.

This patch does not yet support the new scheme for end-to-end footers
on segments during replication.