Remove remaining will code.
Recovery now recovers each tablet on a crashed server to a separate
recovery master and doesn't attempt to double-up tablets. If there
are too few recovery masters the crashed master will be recovered
by a series of recoveries.

Once we have a mechanism in place for tracking the size of tablets
on the coordinator recovery should make an attempt to balance out
work across the recovery masters.