Fix crash with multiple recoveries enqueued for the same server.
Sometimes when two recoveries were scheduled for the same master
(from a hintServerDown and a replacing enlistment, for example)
the second recovery would fail after the first succeeded.
This was because the second recovery couldn't find the master in
the server list.

Now recoveries wait until they are started to find the tablets
which need to be recovered. Because the MasterRecoveryManager
ensures there is only one recovery for a particular master happening
at a time any following recovery for that same server should
find there are no tablets left to recover, in which case its safe
to ignore the fact that the master is no longer in the server list.