Backup reject opens for replicas found on disk after restart.
After restarting a backup finds replicas stored on disk. In the
meantime, a master maybe re-replicating segment which lost replicas.
Since the master is unaware the restarted backup has replicas it
may try to open a replica on that backup.

To prevent confusion and to keep this safe, backups now reject open
requests for replicas if they already have a replica for that segment
which was found on disk after a restart. Open requests to a single
backup process are still idempotent. It is a fatal error to perform any
other write operation on a replica which was found on disk after a
backup restart.